perf htop: Recalculate the column widths for what will be displayed
The common code calculates the column widths for all hist entries,
as 'perf report' needs it as all entries may be navigated, but the
stdio htop shows just what fits in the number of rows in the terminal.

Making it calculate just what appears on the terminal makes the columns
more stable each refresh.

The extra calc is not a problem since we'll right after need those hist
entries in the cache to actually show them.

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>