perf htop: Fix non collapsed splitup
hists__collapse_resort, for the sort__need_collapse case will merge the
existing hist_entries with the new batch, but not when not needing to
collapse, for instance for the default --sort, "dso, sym".

Fix it by introcuding _threaded() variants that will do the merging and
use the right tree.

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>